+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/FHIRPathEngine.java
@@ -263,6 +263,7 @@ public class FHIRPathEngine {
   private boolean allowPolymorphicNames;
   private boolean doImplicitStringConversion;
   private boolean liquidMode; // in liquid mode, || terminates the expression and hands the parser back to the host
+  private boolean doNotEnforceAsSingletonRule;
 
   // if the fhir path expressions are allowed to use constants beyond those defined in the specification
   // the application can implement them by providing a constant resolver 
@@ -454,6 +455,14 @@ public class FHIRPathEngine {
 
   public void setDoImplicitStringConversion(boolean doImplicitStringConversion) {
     this.doImplicitStringConversion = doImplicitStringConversion;
+  }
+
+  public boolean isDoNotEnforceAsSingletonRule() {
+    return doNotEnforceAsSingletonRule;
+  }
+
+  public void setDoNotEnforceAsSingletonRule(boolean doNotEnforceAsSingletonRule) {
+    this.doNotEnforceAsSingletonRule = doNotEnforceAsSingletonRule;
   }
 
   // --- public API -------------------------------------------------------
@@ -1296,14 +1305,14 @@ public class FHIRPathEngine {
 
   private boolean checkParamCount(FHIRLexer lexer, SourceLocation location, ExpressionNode exp, int count) throws FHIRLexerException {
     if (exp.getParameters().size() != count) {
-      throw lexer.error("The function \""+exp.getName()+"\" requires "+Integer.toString(count)+" parameters", location.toString());
+      throw lexer.error("The function \""+exp.getName()+"\" requires "+Integer.toString(count)+" parameters", location.toString(), location);
     }
     return true;
   }
 
   private boolean checkParamCount(FHIRLexer lexer, SourceLocation location, ExpressionNode exp, int countMin, int countMax) throws FHIRLexerException {
     if (exp.getParameters().size() < countMin || exp.getParameters().size() > countMax) {
-      throw lexer.error("The function \""+exp.getName()+"\" requires between "+Integer.toString(countMin)+" and "+Integer.toString(countMax)+" parameters", location.toString());
+      throw lexer.error("The function \""+exp.getName()+"\" requires between "+Integer.toString(countMin)+" and "+Integer.toString(countMax)+" parameters", location.toString(), location);
     }
     return true;
   }
@@ -1788,6 +1797,12 @@ public class FHIRPathEngine {
       return result;
     } else {
       String tn = convertToString(right);
+      if (!isKnownType(tn)) {
+        throw new PathEngineException("The type "+tn+" is not valid");
+      }
+      if (!doNotEnforceAsSingletonRule && left.size() > 1) {
+        throw new PathEngineException("Attempt to use as on more than one item ("+left.size()+")");
+      }
       for (Base nextLeft : left) {
         if (tn.equals(nextLeft.fhirType())) {
           result.add(nextLeft);
@@ -1797,6 +1812,34 @@ public class FHIRPathEngine {
     return result;
   }
 
+
+  private boolean isKnownType(String tn) {
+    if (!tn.contains(".")) {
+      if (Utilities.existsInList(tn, "String", "Boolean", "Integer", "Decimal", "Quantity", "DateTime", "Time", "SimpleTypeInfo", "ClassInfo")) {
+        return true;
+      }
+      try {
+        return worker.fetchTypeDefinition(tn) != null;
+      } catch (Exception e) {
+        return false;
+      }
+    }
+    String[] t = tn.split("\\.");
+    if (t.length != 2) {
+      return false;
+    }
+    if ("System".equals(t[0])) {
+      return Utilities.existsInList(t[1], "String", "Boolean", "Integer", "Decimal", "Quantity", "DateTime", "Time", "SimpleTypeInfo", "ClassInfo");
+    } else if ("FHIR".equals(t[0])) {      
+      try {
+        return worker.fetchTypeDefinition(t[1]) != null;
+      } catch (Exception e) {
+        return false;
+      }
+    } else {
+      return false;
+    }
+  }
 
   private List<Base> opIs(List<Base> left, List<Base> right, ExpressionNode expr) {
     List<Base> result = new ArrayList<Base>();
@@ -3522,7 +3565,7 @@ public class FHIRPathEngine {
     case Aggregate : return funcAggregate(context, focus, exp);
     case Item : return funcItem(context, focus, exp);
     case As : return funcAs(context, focus, exp);
-    case OfType : return funcAs(context, focus, exp);
+    case OfType : return funcOfType(context, focus, exp);
     case Type : return funcType(context, focus, exp);
     case Is : return funcIs(context, focus, exp);
     case Single : return funcSingle(context, focus, exp);
@@ -4552,6 +4595,13 @@ public class FHIRPathEngine {
     } else {
       tn = "FHIR."+expr.getParameters().get(0).getName();
     }
+    if (!isKnownType(tn)) {
+      throw new PathEngineException("The type "+tn+" is not valid");
+    }
+    if (!doNotEnforceAsSingletonRule && focus.size() > 1) {
+      throw new PathEngineException("Attempt to use as() on more than one item ("+focus.size()+")");
+    }
+    
     for (Base b : focus) {
       if (tn.startsWith("System.")) {
         if (b instanceof Element &&((Element) b).isDisallowExtensions()) { 
@@ -4571,7 +4621,48 @@ public class FHIRPathEngine {
               result.add(b);
               break;
             }
-            sd = worker.fetchResource(StructureDefinition.class, sd.getBaseDefinition(), sd);
+            sd = sd.getKind() == StructureDefinitionKind.PRIMITIVETYPE ? null : worker.fetchResource(StructureDefinition.class, sd.getBaseDefinition(), sd);
+          }
+        }
+      }
+    }
+    return result;
+  }
+  
+
+  private List<Base> funcOfType(ExecutionContext context, List<Base> focus, ExpressionNode expr) {
+    List<Base> result = new ArrayList<Base>();
+    String tn;
+    if (expr.getParameters().get(0).getInner() != null) {
+      tn = expr.getParameters().get(0).getName()+"."+expr.getParameters().get(0).getInner().getName();
+    } else {
+      tn = "FHIR."+expr.getParameters().get(0).getName();
+    }
+    if (!isKnownType(tn)) {
+      throw new PathEngineException("The type "+tn+" is not valid");
+    }
+
+    
+    for (Base b : focus) {
+      if (tn.startsWith("System.")) {
+        if (b instanceof Element &&((Element) b).isDisallowExtensions()) { 
+          if (b.hasType(tn.substring(7))) {
+            result.add(b);
+          }
+        }
+
+      } else if (tn.startsWith("FHIR.")) {
+        String tnp = tn.substring(5);
+        if (b.fhirType().equals(tnp)) {
+          result.add(b);          
+        } else {
+          StructureDefinition sd = worker.fetchTypeDefinition(b.fhirType());
+          while (sd != null) {
+            if (tnp.equals(sd.getType())) {
+              result.add(b);
+              break;
+            }
+            sd = sd.getKind() == StructureDefinitionKind.PRIMITIVETYPE ? null : worker.fetchResource(StructureDefinition.class, sd.getBaseDefinition(), sd);
           }
         }
       }