+++ RELEASE_NOTES.md
@@ -1,7 +1,10 @@
 ## Validator Changes
 
-* no changes
+* update FHIRPath implementation for corrections to 'as' (R5 only in validator)
 
 ## Other code changes
 
+* update to latest FHIRPath for older versions 
+* clean up error handling when parsing structure maps
+* go-publish related changes
 * FTP Client upload and logging improvements